+++ packages/config/index.js
@@ -1,8 +1,8 @@
-import {ajaxConfig} from './ajax'
-import {servicesConfig} from './services'
+import { ajaxConfig } from './ajax';
+import { servicesConfig } from './services';
 
-export const ajax = ajaxConfig
-export const services = servicesConfig
+export const ajax = ajaxConfig;
+export const services = servicesConfig;
 
-const configuration = Object.assign({}, ajax, services)
-export default configuration
+const configuration = Object.assign({}, ajax, services);
+export default configuration;