+++ packages/helper/index.js
@@ -1,15 +1,15 @@
 class Helper {    
     checkFormat(value) {
         if (typeof(value)==='number') {
-            let val = (value / 1).toFixed(0).replace(',', '.')
-            return 'Rp ' + val.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ".")
+            let val = (value / 1).toFixed(0).replace(',', '.');
+            return 'Rp ' + val.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ".");
         }
         return value;
         
     }
     formatelipsis(index) {
         if (index.length > 40) {
-            return index.substr(0, 40) + '...'
+            return index.substr(0, 40) + '...';
         }
         return index;
         
@@ -19,4 +19,4 @@ class Helper {
     }
 }
 
-export default Helper
+export default Helper;