+++ webpack.prod.local.js
@@ -34,5 +34,6 @@ config.optimization = Object.assign(config.optimization, {
          }
        }
     }
-})
-module.exports = config
+});
+
+module.exports = config;