+++ webpack.prod.sw.js
@@ -59,6 +59,6 @@ config.optimization = Object.assign(config.optimization, {
          }
        }
     }
-})
+});
 
-module.exports = config
+module.exports = config;